+++ util/opentelemetry-util-genai/CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3891](#3891))
 - Add parent class genAI invocation
   ([https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3889](#3889))
+- Add workflow invocation type to genAI utils
+  ([https://github.com/open-telemetry/opentelemetry-python-contrib/pull/4310](#4310))  
 
 ## Version 0.2b0 (2025-10-14)